Started processing response
