+++ src/cons/state_handler.cpp
@@ -1,7 +1,10 @@
 #include <flatbuffers/flatbuffers.h>
 #include "state_handler.hpp"
 #include "../fbschema/p2pmsg_helpers.hpp"
+#include "../fbschema/p2pmsg_content_generated.h"
 #include "../p2p/p2p.hpp"
+#include "../pchheader.hpp"
+#include "../cons/cons.hpp"
 
 namespace cons
 {
@@ -34,10 +37,45 @@ void handle_state_response()
         {
             std::lock_guard<std::mutex> lock(p2p::ctx.collected_msgs.state_response_mutex);
             candidate_state_responses.clear();
-            p2p::ctx.collected_msgs.state_response.splice(candidate_state_responses);
+            auto it = p2p::ctx.collected_msgs.state_response.begin();
+            candidate_state_responses.splice(candidate_state_responses.end(), p2p::ctx.collected_msgs.state_response, it);
         }
 
+        if (candidate_state_responses.empty())
+            continue;
 
+        if (candidate_state_responses.size() > 1)
+        {
+            LOG_DBG << "Invalid number of state responses to process";
+        }
+
+        for (auto &response : candidate_state_responses)
+        {
+            const fbschema::p2pmsg::Content *content = fbschema::p2pmsg::GetContent(response.data());
+            const fbschema::p2pmsg::State_Response_Message *resp_msg = content->message_as_State_Response_Message();
+
+            const fbschema::p2pmsg::State_Response msg_type = resp_msg->state_response_type();
+            if (msg_type == fbschema::p2pmsg::State_Response_Content_Response)
+            {
+                const fbschema::p2pmsg::Content_Response *con_resp = resp_msg->state_response_as_Content_Response();
+                std::unordered_map<std::string, p2p::state_fs_hash_entry> state_content_list;
+
+                 std::unordered_map<std::string, p2p::state_fs_hash_entry> &&l = fbschema::p2pmsg::flatbuf_statefshashentry_to_statefshashentry(con_resp->content());
+                state_content_list.swap(l);
+
+                std::unordered_map<std::string, p2p::state_fs_hash_entry> existing_fs_entries;
+                bool file_entry_found = false;
+               
+            }
+            else if (msg_type == fbschema::p2pmsg::State_Response_File_HashMap_Response)
+            {
+                const fbschema::p2pmsg::File_HashMap_Response *file_resp = resp_msg->state_response_as_File_HashMap_Response();
+            }
+            else if (msg_type == fbschema::p2pmsg::State_Response_Block_Response)
+            {
+                const fbschema::p2pmsg::Block_Response *block_resp = resp_msg->state_response_as_Block_Response();
+            }
+        }
     }
 }
 } // namespace cons
