+++ src/fbschema/p2pmsg_helpers.hpp
@@ -64,6 +64,15 @@ flatbuf_historyledgermap_to_historyledgermap(const flatbuffers::Vector<flatbuffe
 const flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<HistoryLedgerPair>>>
 historyledgermap_to_flatbuf_historyledgermap(flatbuffers::FlatBufferBuilder &builder, const std::map<uint64_t, const p2p::history_ledger> &map);
 
+std::unordered_map<std::string, p2p::state_fs_hash_entry>
+flatbuf_statefshashentry_to_statefshashentry(const flatbuffers::Vector<flatbuffers::Offset<State_FS_Hash_Entry>> *fhashes);
+
+void statefilehash_to_flatbuf_statefilehash(flatbuffers::FlatBufferBuilder &builder, std::vector<flatbuffers::Offset<State_FS_Hash_Entry>> &list,
+                                            std::string_view full_path, bool is_file, std::string_view hash);
+
+flatbuffers::Offset<Content_Response>
+create_msg_from_content_response(flatbuffers::FlatBufferBuilder &builder, const std::string &fullpath, std::vector<flatbuffers::Offset<State_FS_Hash_Entry>> &content);
+
 } // namespace fbschema::p2pmsg
 
 #endif
